+++ src/engine/Projections.ts
@@ -121,7 +121,7 @@ export const readProjections = async (
     projectionsPath,
     File.readFile,
     Result.mapOk((data: string): string => (data === "" ? "{}" : data)),
-    Result.chainOk((x: string) => File.parseJson<t>(x)),
+    Result.chainOk((x: string) => File.parseJson<t>(x, projectionsPath)),
     Result.mapError((error: string) => ({
       startingFile: projectionsPath,
       message: error
@@ -184,7 +184,7 @@ const mainPathToAlternate = (path: string): string => {
 
   const taggedPath = /\*\*/.test(path) ? path : path.replace("*", "**/*");
 
-  return taggedPath.replace("**", "{dirname}").replace("*", "{basename}");
+  return taggedPath.replace(/\*\*/g, "{dirname}").replace("*", "{basename}");
 };
 
 const alternatePathToAlternate = (path: string): string => {