+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support finding .projections.json in places other than the workspace.
+- Support multiple dirname parts
 
 ### Changed